Move logic to service file and update controllers.
